# SIH/aagaah/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AAGAAH",
    "description": "Explicitly labeled Himalayan replay demonstration.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/pilot": {
      "get": {
        "summary": "Pilot",
        "operationId": "pilot_pilot_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/sources": {
      "get": {
        "summary": "Sources",
        "operationId": "sources_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/model-card": {
      "get": {
        "summary": "Model Card",
        "operationId": "model_card_model_card_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/dashboard": {
      "get": {
        "summary": "Dashboard",
        "operationId": "dashboard_dashboard_get",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "replay",
                "live"
              ],
              "type": "string",
              "default": "replay",
              "title": "Mode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/map/terrain.png": {
      "get": {
        "summary": "Terrain Image",
        "operationId": "terrain_image_map_terrain_png_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/map/{layer}": {
      "get": {
        "summary": "Layer",
        "operationId": "layer_map__layer__get",
        "parameters": [
          {
            "name": "layer",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/locations/{location_id}/history": {
      "get": {
        "summary": "History",
        "operationId": "history_locations__location_id__history_get",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/predict": {
      "post": {
        "summary": "Predict",
        "operationId": "predict_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/replay/control": {
      "post": {
        "summary": "Control",
        "operationId": "control_replay_control_post",
        "parameters": [
          {
            "name": "x-control-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Control-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayControl"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/tick": {
      "post": {
        "summary": "Scheduled Tick",
        "operationId": "scheduled_tick_internal_tick_post",
        "parameters": [
          {
            "name": "x-control-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Control-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PredictRequest": {
        "properties": {
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "readings": {
            "items": {
              "$ref": "#/components/schemas/Reading"
            },
            "type": "array",
            "maxItems": 10000,
            "minItems": 1,
            "title": "Readings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "as_of",
          "readings"
        ],
        "title": "PredictRequest"
      },
      "Reading": {
        "properties": {
          "location_id": {
            "type": "string",
            "title": "Location Id"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "available_at": {
            "type": "string",
            "format": "date-time",
            "title": "Available At"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "synthetic": {
            "type": "boolean",
            "title": "Synthetic",
            "default": false
          },
          "rainfall_mm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 500.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rainfall Mm"
          },
          "soil_moisture": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Soil Moisture"
          },
          "water_level_m": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 50.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Water Level M"
          },
          "forecast_rain_mm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 500.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Rain Mm"
          },
          "forecast_issued_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Issued At"
          },
          "interval_minutes": {
            "type": "integer",
            "maximum": 1440.0,
            "minimum": 1.0,
            "title": "Interval Minutes",
            "default": 60
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "location_id",
          "observed_at",
          "available_at",
          "source",
          "status"
        ],
        "title": "Reading"
      },
      "ReplayControl": {
        "properties": {
          "action": {
            "type": "string",
            "pattern": "^(play|pause|step|reset|seek)$",
            "title": "Action"
          },
          "step": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 71.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Step"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "ReplayControl"
      },
      "Status": {
        "type": "string",
        "enum": [
          "VERIFIED LIVE",
          "MOCKED",
          "REPLAY",
          "PLANNED"
        ],
        "title": "Status"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}